Beta updates
Adds the Beta context to the README. Adds rubocop to the default rake
task, as well as the CI pipeline. Renames the openapi spec to
mx_platform_api.yml and moves it to an `/openapi` folder. This allows
for multiple openapi specs in the future. Updates the Rakefile with
rubocop changes, as well as refactor to handle multiple specs.