Consolidate tags
Consolidates tags into a single `mx_platform` tag. This was done because
the openapi-generator will treat each tag as it's own "api". This would
require instantiating multiple "apis" in the library to access
the endpoints that belonged to that tag. By consolidating tags into a
single tag, we now have all the endpoints under a single "api".