Refactor test workflow
Previously with our test workflow we were running a rake task that would
validate the YAML in the openapi spec and we would validate the openapi
schema to ensure it followed the openapi specification.

We did this using two gems, the `yaml_normalizer` gem and the
`openapi3_parser` gem. The problem is that both gems had the `psych` gem
pinned which caused dependency issues. As a workaround we forked the
`yaml_normalizer` gem and unpinned the `psych` gem. This worked but has
the potential for issues and could cause difficulties when
troubleshooting workflow problems.

One thing that was discovered recently, is that the `openapi3_parser`
gem wasn't picking up any errors from schema when there were errors
present. So in an effort to simplify our workflows and to not rely on
any brittle workarounds, lets use some GitHub actions that have been
built for this very purpose.

We will still make use of the `yaml_normalizer` gem to normalize the
file locally but we will rely on the GitHub workflows to validate and
check for errors.