P1: Autogenerate client libraries on OpenAPI changes
This PR sets the base for being able to automate generating and publishing our client libraries from end-to-end based on changes made to the OpenAPI spec.

We will do this with a `version` action that first checks for changes made to the openapi spec and if so, we will then check for a `major`, `minor`, or `patch` label so that we can determine how to update the version for our libraries.