P2: Autogenerate libraries
This builds off of [this PR](https://github.com/mxenabled/openapi/pull/100) to autogenerate our client libraries.

We will do this by utilizing a GitHub App so that we can kick off actions in the correct repos to generate and update our libraries when changes are made to the OpenAPI spec.